+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Enhancements
 * Variable definitions are optimized so that shadowed variables are reused as much as possible to reduce memory consumption.
 * `FnAccess` and `FnNamespace` now implement `Ord` and `PartialOrd`.
 * The `event_handler_map` example is enhanced to prevent shadowing of the state object map.
+* Separation of constants in function calls is removed as its performance benefit is dubious.
 
 
 Version 1.5.0